:recycle: rename to WindowControl from WindowControlMenu
