+++ src/domain/window/views/WindoeControl/WindowControl.css.ts
