+++ src/domain/window/views/WindoeControl/WindowControl.tsx
@@ -1,0 +1,9 @@
+import { container } from './WindowControl.css.ts';
+
+export function WindowControl() {
+  return (
+    <div className={container}>
+      <p>WindowControlMenu</p>
+    </div>
+  );
+}
