+++ src/domain/window/views/WindoeControl/index.ts
@@ -1,0 +1,1 @@
+export * from './WindowControl.tsx';
