--- src/domain/window/views/WindoeControlMenu/WindowControlMenu.tsx
@@ -1,9 +1,0 @@
-import { container } from './WindowControlMenu.css.ts';
-
-export function WindowControlMenu() {
-  return (
-    <div className={container}>
-      <p>WindowControlMenu</p>
-    </div>
-  );
-}
