--- src/domain/window/views/WindoeControlMenu/index.ts
@@ -1,1 +1,0 @@
-export * from './WindowControlMenu.tsx';
