+++ src/domain/window/views/Window/Window.tsx
@@ -1,5 +1,5 @@
 import { container } from './Window.css.ts';
-import { WindowControlMenu } from '../WindoeControlMenu';
+import { WindowControl } from '../WindoeControl';
 
 interface WindowProps {
   children?: React.ReactNode;
@@ -9,4 +9,4 @@ export function Window({ children }: WindowProps) {
   return <div className={container}>{children}</div>;
 }
 
-Window.ControlMenu = WindowControlMenu;
+Window.Control = WindowControl;
